Move misc packages into pkg package
This is one of the main bits of feedback from @StefanProdan
that was still remaining after consolidating the two
controllers.

This patch also removes the service namespace field from
the scale type. The value can and is read from the query
string, meaning that this field was redundant and non-standard
compared with the types in faas-provider.

Tested with go test ./... and with make using the Dockerfile.
