+++ pkg/config/read_config.go
@@ -1,6 +1,7 @@
 // Copyright (c) Alex Ellis 2017. All rights reserved.
+// Copyright (c) OpenFaaS Author(s) 2020. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-package types
+package config
 
 import (
 	"fmt"
