+++ pkg/handlers/delete.go
