+++ pkg/handlers/secrets.go
@@ -8,7 +8,7 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/openfaas/faas-netes/k8s"
+	"github.com/openfaas/faas-netes/pkg/k8s"
 	types "github.com/openfaas/faas-provider/types"
 	"k8s.io/client-go/kubernetes"
 )
